# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminCredentials {
  password: string;
}

export interface AuthResult {
  ok: boolean;
}

export interface ProfileData {
  id: number;
  name: string;
  title: string;
  bio: string;
  location: string;
  email: string;
  phone: string;
  linkedinUrl: string;
  githubUrl: string;
  websiteUrl: string;
  avatarUrl: string;
  updatedAt: string;
}

export interface ProfileInput {
  name?: string;
  title?: string;
  bio?: string;
  location?: string;
  email?: string;
  phone?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  websiteUrl?: string;
  avatarUrl?: string;
}

export interface ArticleData {
  id: number;
  slug: string;
  title: string;
  subtitle: string;
  content: string;
  publishedDate: string;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ArticleInput {
  slug: string;
  title: string;
  subtitle?: string;
  content?: string;
  publishedDate: string;
  published?: boolean;
}

export interface ProjectData {
  id: number;
  name: string;
  description: string;
  url: string;
  tags: string[];
  displayOrder: number;
  createdAt: string;
}

export interface ProjectInput {
  name: string;
  description?: string;
  url?: string;
  tags?: string[];
  displayOrder?: number;
}

export interface ExperienceData {
  id: number;
  company: string;
  companyUrl: string;
  role: string;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  description: string;
  displayOrder: number;
  createdAt: string;
}

export interface ExperienceInput {
  company: string;
  companyUrl?: string;
  role: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  description?: string;
  displayOrder?: number;
}

export interface EducationData {
  id: number;
  institution: string;
  degree: string;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  description: string;
  displayOrder: number;
  createdAt: string;
}

export interface EducationInput {
  institution: string;
  degree: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  description?: string;
  displayOrder?: number;
}

export type ListArticlesParams = {
includeUnpublished?: string;
};