# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: Profile operations
  - name: articles
    description: Article operations
  - name: projects
    description: Project operations
  - name: experience
    description: Experience operations
  - name: education
    description: Education operations
  - name: auth
    description: Auth operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/verify:
    post:
      operationId: verifyAdmin
      tags: [auth]
      summary: Verify admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get profile
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileData"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileData"
        "401":
          description: Unauthorized

  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List all articles
      parameters:
        - name: includeUnpublished
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleData"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleData"
        "401":
          description: Unauthorized

  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleData"
        "404":
          description: Not found
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update article (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleData"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectData"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectData"
        "401":
          description: Unauthorized

  /projects/{id}:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectData"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /experience:
    get:
      operationId: listExperience
      tags: [experience]
      summary: List experience
      responses:
        "200":
          description: Experience list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExperienceData"
    post:
      operationId: createExperience
      tags: [experience]
      summary: Create experience entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperienceData"
        "401":
          description: Unauthorized

  /experience/{id}:
    put:
      operationId: updateExperience
      tags: [experience]
      summary: Update experience entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperienceData"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteExperience
      tags: [experience]
      summary: Delete experience entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /education:
    get:
      operationId: listEducation
      tags: [education]
      summary: List education
      responses:
        "200":
          description: Education list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EducationData"
    post:
      operationId: createEducation
      tags: [education]
      summary: Create education entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationData"
        "401":
          description: Unauthorized

  /education/{id}:
    put:
      operationId: updateEducation
      tags: [education]
      summary: Update education entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationData"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteEducation
      tags: [education]
      summary: Delete education entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AdminCredentials:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AuthResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    ProfileData:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        location:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        websiteUrl:
          type: string
        avatarUrl:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - title
        - bio
        - location
        - email
        - phone
        - linkedinUrl
        - githubUrl
        - websiteUrl
        - avatarUrl
        - updatedAt

    ProfileInput:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        location:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        websiteUrl:
          type: string
        avatarUrl:
          type: string

    ArticleData:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
        content:
          type: string
        publishedDate:
          type: string
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - title
        - subtitle
        - content
        - publishedDate
        - published
        - createdAt
        - updatedAt

    ArticleInput:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
        content:
          type: string
        publishedDate:
          type: string
        published:
          type: boolean
      required:
        - slug
        - title
        - publishedDate

    ProjectData:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        url:
          type: string
        tags:
          type: array
          items:
            type: string
        displayOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - url
        - tags
        - displayOrder
        - createdAt

    ProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
        tags:
          type: array
          items:
            type: string
        displayOrder:
          type: integer
      required:
        - name

    ExperienceData:
      type: object
      properties:
        id:
          type: integer
        company:
          type: string
        companyUrl:
          type: string
        role:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        description:
          type: string
        displayOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - company
        - companyUrl
        - role
        - startDate
        - endDate
        - description
        - displayOrder
        - createdAt

    ExperienceInput:
      type: object
      properties:
        company:
          type: string
        companyUrl:
          type: string
        role:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        description:
          type: string
        displayOrder:
          type: integer
      required:
        - company
        - role
        - startDate

    EducationData:
      type: object
      properties:
        id:
          type: integer
        institution:
          type: string
        degree:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        description:
          type: string
        displayOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - institution
        - degree
        - startDate
        - endDate
        - description
        - displayOrder
        - createdAt

    EducationInput:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        description:
          type: string
        displayOrder:
          type: integer
      required:
        - institution
        - degree
        - startDate