# lib/api-zod/src/generated/types/articleData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleData {
  id: number;
  slug: string;
  title: string;
  subtitle: string;
  content: string;
  publishedDate: string;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}