# lib/api-zod/src/generated/types/educationData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EducationData {
  id: number;
  institution: string;
  degree: string;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  description: string;
  displayOrder: number;
  createdAt: string;
}