# lib/api-zod/src/generated/types/profileData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileData {
  id: number;
  name: string;
  title: string;
  bio: string;
  location: string;
  email: string;
  phone: string;
  linkedinUrl: string;
  githubUrl: string;
  websiteUrl: string;
  avatarUrl: string;
  updatedAt: string;
}