# lib/api-zod/src/generated/types/projectData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectData {
  id: number;
  name: string;
  description: string;
  url: string;
  tags: string[];
  displayOrder: number;
  createdAt: string;
}